fix(update): improve the partner management endpoint

+++ src/router/v3/partners/manage.js
@@ -138,43 +138,58 @@ module.exports = async function (fastify) {
             const { title, image, bio, url, social, Method } = request.body
 
             if (Method === 'add') {
-                await sqlQuery({
-                    query: `INSERT INTO partners (title, image, bio, url, social) VALUES ("${title}", "${image}", "${bio}", "${url}", "${social}")`
-                })
-                    .then(() => {
+                const partner = await sqlQuery({ query: `SELECT * FROM partners WHERE title="${title}"` }).then(p => p)
+
+                await console.log(partner[0])
+
+                if (partner[0]) {
+                    return reply.code(400).send({
+                        code: 'PARTNER_EXISTS',
+                        message: 'We were unable to add that partner',
+                        errormsg: 'partner already exists in our system',
+                        status: 400
+                    })
+                } else {
+                    await sqlQuery({
+                        query: `INSERT INTO partners (title, image, bio, url, social) VALUES ("${title}", "${image}", "${bio}", "${url}", "${social}")`
+                    }).then(() => {
                         return reply.code(200).send({
                             code: 'SUCCESS',
                             message: 'Partner has been added',
                             status: 200
                         })
                     })
-                    .catch(e => {
-                        return reply.code(500).send({
-                            code: 'ERROR',
-                            message: 'Failed to add partner',
-                            errormsg: `${e.message}`,
-                            status: 500
-                        })
-                    })
+                }
             } else if (Method === 'delete') {
-                await sqlQuery({
-                    query: `DELETE FROM partners WHERE title = "${title}"`
-                })
-                    .then(() => {
-                        return reply.code(200).send({
-                            code: 'SUCCESS',
-                            message: 'Partner has been deleted',
-                            status: 200
-                        })
+                const partner = await sqlQuery({ query: `SELECT * FROM partners WHERE title="${title}"` })
+
+                if (!partner[0]) {
+                    return reply.code(404).send({
+                        code: 'PARTNER_NOT_FOUND',
+                        message: 'We were unable to locate that partner',
+                        errormsg: 'are you sure a partner with this name exists?',
+                        status: 404
                     })
-                    .catch(e => {
-                        return reply.code(500).send({
-                            code: 'ERROR',
-                            message: 'Failed to delete partner',
-                            errormsg: `${e.message}`,
-                            status: 500
-                        })
+                } else {
+                    await sqlQuery({
+                        query: `DELETE FROM partners WHERE title = "${title}"`
                     })
+                        .then(() => {
+                            return reply.code(200).send({
+                                code: 'SUCCESS',
+                                message: 'Partner has been deleted',
+                                status: 200
+                            })
+                        })
+                        .catch(e => {
+                            return reply.code(500).send({
+                                code: 'ERROR',
+                                message: 'Failed to delete partner',
+                                errormsg: `${e.message}`,
+                                status: 500
+                            })
+                        })
+                }
             } else {
                 return reply.code(400).send({
                     code: 'MALFORMED_REQUEST',
